switching to int type for worker task constants and removing magic number from worker unit test

+++ MRBrew/MRBrewWorkerTaskConstants.h
@@ -25,5 +25,5 @@
 
 #import <Foundation/Foundation.h>
 
-extern NSUInteger const MRBrewWorkerTaskExitedNormally;
-extern NSUInteger const MRBrewWorkerTaskCancelled;
+extern int const MRBrewWorkerTaskExitedNormally;
+extern int const MRBrewWorkerTaskCancelled;
